Reuse some of the shared code between nodes
