+++ beluga_amcl/src/ndt_amcl_node.cpp
@@ -1,4 +1,4 @@
-// Copyright 2022-2024 Ekumen, Inc.
+// Copyright 2024 Ekumen, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,25 +21,23 @@
 #include <beluga/random/multivariate_normal_distribution.hpp>
 #include <beluga/sensor/ndt_sensor_model.hpp>
 #include <beluga/views/particles.hpp>
+#include <beluga_amcl/common.hpp>
 #include <beluga_amcl/ndt_amcl_node.hpp>
 
 #include <tf2/convert.h>
 #include <tf2/utils.h>
 #include <tf2_ros/create_timer_ros.h>
 
-#include <beluga_ros/amcl.hpp>
 #include <chrono>
 #include <execution>
-#include <limits>
 #include <memory>
 #include <sophus/se2.hpp>
 #include <stdexcept>
 #include <string>
 #include <string_view>
 #include <tuple>
-#include <type_traits>
-#include <unordered_map>
 #include <utility>
+#include <variant>
 
 #include <geometry_msgs/msg/pose_stamped.hpp>
 #include <lifecycle_msgs/msg/state.hpp>
@@ -47,282 +45,19 @@
 #include <range/v3/range/conversion.hpp>
 #include <rclcpp_components/register_node_macro.hpp>
 
+#include <beluga_ros/amcl.hpp>
 #include <beluga_ros/tf2_sophus.hpp>
-#include <variant>
 
 namespace beluga_amcl {
-
-namespace {
-
-constexpr std::string_view kDifferentialModelName = "differential_drive";
-constexpr std::string_view kOmnidirectionalModelName = "omnidirectional_drive";
-constexpr std::string_view kStationaryModelName = "stationary";
-constexpr std::string_view kNav2DifferentialModelName = "nav2_amcl::DifferentialMotionModel";
-constexpr std::string_view kNav2OmnidirectionalModelName = "nav2_amcl::OmniMotionModel";
-
-}  // namespace
 
 NdtAmclNode::NdtAmclNode(const rclcpp::NodeOptions& options)
     : rclcpp_lifecycle::LifecycleNode{"ndt_amcl", "", options} {
   RCLCPP_INFO(get_logger(), "Creating");
 
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "The name of the coordinate frame published by the localization system.";
-    declare_parameter("global_frame_id", rclcpp::ParameterValue("map"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "The name of the coordinate frame to use for odometry.";
-    declare_parameter("odom_frame_id", rclcpp::ParameterValue("odom"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "The name of the coordinate frame to use for the robot base.";
-    declare_parameter("base_frame_id", rclcpp::ParameterValue("base_footprint"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Path to load the map from an hdf5 file.";
-    declare_parameter("map_path", rclcpp::ParameterValue("map_path"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Topic to subscribe to in order to receive the initial pose of the robot.";
-    declare_parameter("initial_pose_topic", rclcpp::ParameterValue("initialpose"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Topic to subscribe to in order to receive the laser scan for localization.";
-    declare_parameter("scan_topic", rclcpp::ParameterValue("scan"), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Minimum allowed number of particles.";
-    descriptor.integer_range.resize(1);
-    descriptor.integer_range[0].from_value = 0;
-    descriptor.integer_range[0].to_value = std::numeric_limits<int>::max();
-    descriptor.integer_range[0].step = 1;
-    declare_parameter("min_particles", rclcpp::ParameterValue(500), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Maximum allowed number of particles.";
-    descriptor.integer_range.resize(1);
-    descriptor.integer_range[0].from_value = 0;
-    descriptor.integer_range[0].to_value = std::numeric_limits<int>::max();
-    descriptor.integer_range[0].step = 1;
-    declare_parameter("max_particles", rclcpp::ParameterValue(2000), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Exponential decay rate for the slow average weight filter, used in deciding when to recover "
-        "by adding random poses.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = 1;
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("recovery_alpha_slow", rclcpp::ParameterValue(0.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Exponential decay rate for the fast average weight filter, used in deciding when to recover "
-        "by adding random poses.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = 1;
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("recovery_alpha_fast", rclcpp::ParameterValue(0.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Maximum particle filter population error between the true distribution "
-        "and the estimated distribution. It is used in KLD resampling to limit the "
-        "allowed number of particles to the minimum necessary.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = 1;
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("pf_err", rclcpp::ParameterValue(0.05), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Upper standard normal quantile for P, where P is the probability "
-        "that the error in the estimated distribution will be less than pf_err "
-        "in KLD resampling.";
-    declare_parameter("pf_z", rclcpp::ParameterValue(0.99), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Resolution in meters for the X axis used to divide the space in buckets for KLD resampling.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("spatial_resolution_x", rclcpp::ParameterValue(0.5), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Resolution in meters for the Y axis used to divide the space in buckets for KLD resampling.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("spatial_resolution_y", rclcpp::ParameterValue(0.5), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Resolution in radians for the theta axis to divide the space in buckets for KLD resampling.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = 2 * Sophus::Constants<double>::pi();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter(
-        "spatial_resolution_theta", rclcpp::ParameterValue(10 * Sophus::Constants<double>::pi() / 180), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Number of filter updates required before resampling. ";
-    descriptor.integer_range.resize(1);
-    descriptor.integer_range[0].from_value = 1;
-    descriptor.integer_range[0].to_value = std::numeric_limits<int>::max();
-    descriptor.integer_range[0].step = 1;
-    declare_parameter("resample_interval", rclcpp::ParameterValue(1), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "When set to true, will reduce the resampling rate when not needed and help "
-        "avoid particle deprivation. The resampling will only happen if the effective "
-        "number of particles (N_eff = 1/(sum(k_i^2))) is lower than half the current "
-        "number of particles.";
-    descriptor.read_only = true;
-    declare_parameter("selective_resampling", false, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Set this to false to prevent amcl from publishing the transform "
-        "between the global frame and the odometry frame.";
-    declare_parameter("tf_broadcast", rclcpp::ParameterValue(true), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description =
-        "Time with which to post-date the transform that is published, "
-        "to indicate that this transform is valid into the future";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("transform_tolerance", rclcpp::ParameterValue(1.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Which motion model to use [differential_drive, omnidirectional_drive, stationary].";
-    declare_parameter("robot_model_type", rclcpp::ParameterValue(std::string(kDifferentialModelName)), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Rotation noise from rotation for the differential drive model.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("alpha1", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Rotation noise from translation for the differential drive model.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("alpha2", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Translation noise from translation for the differential drive model.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("alpha3", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Translation noise from rotation for the differential drive model.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("alpha4", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Strafe noise from translation for the omnidirectional drive model.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("alpha5", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Rotational movement required before performing a filter update.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = 2 * Sophus::Constants<double>::pi();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("update_min_a", rclcpp::ParameterValue(0.2), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Translational movement required before performing a filter update.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("update_min_d", rclcpp::ParameterValue(0.25), descriptor);
-  }
-
+  declare_common_params(*this);
   {
     auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
     descriptor.description = "Likelihood for measurements that lie inside cells that are not present in the map.";
-    // TODO(alon): check if the description values are fine.
     descriptor.floating_point_range.resize(1);
     descriptor.floating_point_range[0].from_value = 0;
     descriptor.floating_point_range[0].to_value = 1;
@@ -333,7 +68,6 @@ NdtAmclNode::NdtAmclNode(const rclcpp::NodeOptions& options)
   {
     auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
     descriptor.description = "Scaling parameter d1 in literature, used for scaling 2D likelihoods.";
-    // TODO(alon): check if the description values are fine.
     descriptor.floating_point_range.resize(1);
     descriptor.floating_point_range[0].from_value = 0;
     descriptor.floating_point_range[0].to_value = 1000;
@@ -344,108 +78,11 @@ NdtAmclNode::NdtAmclNode(const rclcpp::NodeOptions& options)
   {
     auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
     descriptor.description = "Scaling parameter d2 in literature, used for scaling 2D likelihoods.";
-    // TODO(alon): check if the description values are fine.
     descriptor.floating_point_range.resize(1);
     descriptor.floating_point_range[0].from_value = 0;
     descriptor.floating_point_range[0].to_value = 1000;
     descriptor.floating_point_range[0].step = 0;
     declare_parameter("d2", rclcpp::ParameterValue(1.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Maximum scan range to be considered.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("laser_max_range", rclcpp::ParameterValue(100.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Minimum scan range to be considered.";
-    descriptor.floating_point_range.resize(1);
-    descriptor.floating_point_range[0].from_value = 0;
-    descriptor.floating_point_range[0].to_value = std::numeric_limits<double>::max();
-    descriptor.floating_point_range[0].step = 0;
-    declare_parameter("laser_min_range", rclcpp::ParameterValue(0.0), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "How many evenly-spaced beams in each scan will be used when updating the filter.";
-    descriptor.integer_range.resize(1);
-    descriptor.integer_range[0].from_value = 2;
-    descriptor.integer_range[0].to_value = std::numeric_limits<int>::max();
-    descriptor.integer_range[0].step = 1;
-    declare_parameter("max_beams", rclcpp::ParameterValue(60), descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Set the initial pose from the initial_pose parameters.";
-    declare_parameter("set_initial_pose", false, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose x axis coordinate.";
-    declare_parameter("initial_pose.x", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose y axis coordinate.";
-    declare_parameter("initial_pose.y", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose yaw rotation.";
-    declare_parameter("initial_pose.yaw", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose x axis covariance.";
-    declare_parameter("initial_pose.covariance_x", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose y axis covariance.";
-    declare_parameter("initial_pose.covariance_y", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose yaw covariance.";
-    declare_parameter("initial_pose.covariance_yaw", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose xy covariance.";
-    declare_parameter("initial_pose.covariance_xy", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose xyaw covariance.";
-    declare_parameter("initial_pose.covariance_xyaw", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Initial pose yyaw covariance.";
-    declare_parameter("initial_pose.covariance_yyaw", 0.0, descriptor);
-  }
-
-  {
-    auto descriptor = rcl_interfaces::msg::ParameterDescriptor();
-    descriptor.description = "Execution policy used to process particles [seq, par].";
-    declare_parameter("execution_policy", "seq", descriptor);
   }
 }
 
@@ -662,9 +299,10 @@ auto NdtAmclNode::make_particle_filter() const -> std::unique_ptr<NdtAmclVariant
 
         RandomStateGenerator random_state_maker = [](const auto& particles) {
           static thread_local auto generator = std::mt19937{std::random_device()()};
-          const auto [pose, covariance] =
-              beluga::estimate(beluga::views::states(particles), beluga::views::weights(particles));
-          return [pose, covariance]() { return beluga::MultivariateNormalDistribution{pose, covariance}(generator); };
+          const auto estimate = beluga::estimate(beluga::views::states(particles), beluga::views::weights(particles));
+          return [estimate]() {
+            return beluga::MultivariateNormalDistribution{estimate.first, estimate.second}(generator);
+          };
         };
 
         return beluga::Amcl(
@@ -687,42 +325,6 @@ void NdtAmclNode::timer_callback() {
   if (particle_cloud_pub_->get_subscription_count() == 0) {
     return;
   }
-
-  // Particle weights from the filter may or may not be representative of the
-  // true distribution. If we resampled, they are not, and there will be multiple copies
-  // of the most likely candidates, all with unit weight. In this case the number of copies
-  // is a proxy for the prob density at each candidate. If we did not resample before updating
-  // the estimation and publishing this message (which can happen if the resample interval
-  // is set to something other than 1), then all particles are expected to be different
-  // and their weights are proportional to the prob density at each candidate.
-  //
-  // Only the combination of both the state distribution and the candidate weights together
-  // provide information about the probability density at each candidate.
-  // To handle both cases, we group repeated candidates and compute the accumulated weight
-
-  struct RepresentativeData {
-    Sophus::SE2d state;
-    double weight{0.};
-  };
-
-  struct RepresentativeBinHash {
-    std::size_t operator()(const Sophus::SE2d& s) const noexcept {
-      std::size_t h1 = std::hash<double>{}(s.translation().x());
-      std::size_t h2 = std::hash<double>{}(s.translation().y());
-      std::size_t h3 = std::hash<double>{}(s.so2().log());
-      return h1 ^ (h2 << 1) ^ (h3 << 2);
-    }
-  };
-
-  struct RepresentativeBinEqual {
-    bool operator()(const Sophus::SE2d& lhs, const Sophus::SE2d& rhs) const noexcept {
-      // good enough, since copies of the same candidate are expected to be identical copies
-      return lhs.translation().x() == rhs.translation().x() &&  //
-             lhs.translation().y() == rhs.translation().y() &&  //
-             lhs.so2().log() == rhs.so2().log();
-    }
-  };
-
   std::visit(
       [this](const auto& particle_filter) {
         std::unordered_map<Sophus::SE2d, RepresentativeData, RepresentativeBinHash, RepresentativeBinEqual>
@@ -888,11 +490,9 @@ bool NdtAmclNode::initialize_from_estimate(const std::pair<Sophus::SE2d, Eigen::
     return false;
   }
 
-  const auto& [pose, covariance] = estimate;
-
   try {
     std::visit(
-        [&pose, &covariance](auto& particle_filter) { particle_filter.initialize(pose, covariance); },
+        [estimate](auto& particle_filter) { particle_filter.initialize(estimate.first, estimate.second); },
         *particle_filter_);
   } catch (const std::runtime_error& error) {
     RCLCPP_ERROR(get_logger(), "Could not initialize particles: %s", error.what());
@@ -904,6 +504,7 @@ bool NdtAmclNode::initialize_from_estimate(const std::pair<Sophus::SE2d, Eigen::
   const auto num_particles =
       std::visit([](const auto& particle_filter) { return particle_filter.particles().size(); }, *particle_filter_);
 
+  const auto& pose = estimate.first;
   RCLCPP_INFO(
       get_logger(), "Particle filter initialized with %ld particles about initial pose x=%g, y=%g, yaw=%g",
       num_particles, pose.translation().x(), pose.translation().y(), pose.so2().log());
