+++ beluga_amcl/test/test_ndt_amcl_node.cpp
@@ -1,4 +1,4 @@
-// Copyright 2023 Ekumen, Inc.
+// Copyright 2024 Ekumen, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,67 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <beluga_amcl/ndt_amcl_node.hpp>
+
+#include "test_utils/node_testing.hpp"
+
 #include <gmock/gmock.h>
+
 #include <tf2_ros/create_timer_ros.h>
 #include <ament_index_cpp/get_package_prefix.hpp>
-
-#include <beluga_amcl/ndt_amcl_node.hpp>
-#include <beluga_ros/tf2_sophus.hpp>
+#include <lifecycle_msgs/msg/state.hpp>
 
 #include <cstdlib>
-#include <iostream>
-#include <lifecycle_msgs/msg/state.hpp>
+
+#include <beluga_ros/tf2_sophus.hpp>
 
 namespace {
 
 const std::string kTestMapPath = "./test_data/turtlebot3_world.hdf5";
 
 using namespace std::chrono_literals;
-
-/// Spin a group of nodes until a condition is met.
-/**
- * \param predicate The stop condition.
- * \param timeout Maximum time to spin.
- * \param nodes The nodes to spin.
- * \return True if the condition was met. False if it timed out.
- */
-template <class Predicate, class Rep, class Period, class... Nodes>
-bool spin_until(
-    Predicate&& predicate,
-    const std::chrono::duration<Rep, Period>& timeout,
-    const std::shared_ptr<Nodes>&... nodes) {
-  rclcpp::executors::SingleThreadedExecutor executor;
-  (executor.add_node(nodes->get_node_base_interface()), ...);
-
-  const auto deadline = std::chrono::high_resolution_clock::now() + timeout;
-  while (rclcpp::ok() && !predicate() && std::chrono::high_resolution_clock::now() < deadline) {
-    executor.spin_once(deadline - std::chrono::high_resolution_clock::now());  // wait for it
-    executor.spin_some(deadline - std::chrono::high_resolution_clock::now());  // flush it all out
-  }
-  return predicate();  // last minute check
-}
-
-/// Spin a group of nodes until a condition is met with a default timeout.
-/**
- * \param predicate The stop condition.
- * \param nodes The nodes to spin.
- * \return True if the condition was met. False if it timed out.
- */
-template <class Predicate, class... Nodes>
-bool spin_until(Predicate&& predicate, const std::shared_ptr<Nodes>&... nodes) {
-  return spin_until(std::forward<Predicate>(predicate), 10s, nodes...);
-}
-
-/// Spin a group of nodes for a given duration of time.
-/**
- * \param predicate The stop condition.
- * \param duration Time to spin.
- */
-template <class Rep, class Period, class... Nodes>
-void spin_for(const std::chrono::duration<Rep, Period>& duration, const std::shared_ptr<Nodes>&... nodes) {
-  const auto duration_is_over = []() { return false; };
-  spin_until(duration_is_over, duration, nodes...);
-}
+using beluga_amcl::testing::spin_for;
+using beluga_amcl::testing::spin_until;
 
 /// Test class that provides convenient public accessors.
 class AmclNodeUnderTest : public beluga_amcl::NdtAmclNode {
@@ -87,177 +47,6 @@ class AmclNodeUnderTest : public beluga_amcl::NdtAmclNode {
   const auto& estimate() { return last_known_estimate_.value(); }
 };
 
-// Tester node that can publish default messages and test ROS interactions.
-class TesterNode : public rclcpp::Node {
- public:
-  TesterNode() : rclcpp::Node{"tester_node", "", rclcpp::NodeOptions()} {
-    initial_pose_publisher_ =
-        create_publisher<geometry_msgs::msg::PoseWithCovarianceStamped>("initialpose", rclcpp::SystemDefaultsQoS());
-
-    laser_scan_publisher_ = create_publisher<sensor_msgs::msg::LaserScan>("scan", rclcpp::SystemDefaultsQoS());
-  }
-
-  void create_transform_buffer() {
-    // NOTE(nahuel): This cannot be called in the constructor as it uses shared_from_this().
-    // That's why we provide an initialization method.
-    tf_buffer_ = std::make_unique<tf2_ros::Buffer>(get_clock());
-    tf_buffer_->setCreateTimerInterface(
-        std::make_shared<tf2_ros::CreateTimerROS>(get_node_base_interface(), get_node_timers_interface()));
-    tf_broadcaster_ = std::make_unique<tf2_ros::TransformBroadcaster>(shared_from_this());
-    tf_listener_ = std::make_unique<tf2_ros::TransformListener>(
-        *tf_buffer_, this,
-        false);  // avoid using dedicated tf thread
-  }
-
-  void create_pose_subscriber() {
-    pose_subscriber_ = create_subscription<geometry_msgs::msg::PoseWithCovarianceStamped>(
-        "pose", rclcpp::SystemDefaultsQoS(), std::bind(&TesterNode::pose_callback, this, std::placeholders::_1));
-  }
-
-  void pose_callback(geometry_msgs::msg::PoseWithCovarianceStamped::SharedPtr message) { latest_pose_ = *message; }
-
-  auto& latest_pose() { return latest_pose_; }
-
-  void create_particle_cloud_subscriber() {
-    particle_cloud_subscriber_ = create_subscription<nav2_msgs::msg::ParticleCloud>(
-        "particle_cloud", rclcpp::SystemDefaultsQoS(),
-        std::bind(&TesterNode::particle_cloud_callback, this, std::placeholders::_1));
-  }
-
-  void particle_cloud_callback(nav2_msgs::msg::ParticleCloud::SharedPtr message) { latest_particle_cloud_ = *message; }
-
-  const auto& latest_particle_cloud() const { return latest_particle_cloud_; }
-
-  static auto make_dummy_map() {
-    auto map = nav_msgs::msg::OccupancyGrid{};
-    map.header.frame_id = "map";
-    map.info.resolution = 1.0;
-    map.info.width = 2;
-    map.info.height = 2;
-    map.data = std::vector<std::int8_t>{0, 0, 0, 0};
-    return map;
-  }
-
-  void publish_default_initial_pose() {
-    auto pose = geometry_msgs::msg::PoseWithCovarianceStamped{};
-    pose.header.frame_id = "map";
-    initial_pose_publisher_->publish(pose);
-  }
-
-  void publish_initial_pose(double x, double y) {
-    auto pose = geometry_msgs::msg::PoseWithCovarianceStamped{};
-    pose.header.frame_id = "map";
-    pose.pose.pose.position.x = x;
-    pose.pose.pose.position.y = y;
-    initial_pose_publisher_->publish(pose);
-  }
-
-  void publish_initial_pose_with_wrong_frame() {
-    auto pose = geometry_msgs::msg::PoseWithCovarianceStamped{};
-    pose.header.frame_id = "non_existing_frame";
-    initial_pose_publisher_->publish(pose);
-  }
-
-  void publish_laser_scan() {
-    const auto timestamp = now();
-
-    auto scan = sensor_msgs::msg::LaserScan{};
-    scan.header.stamp = timestamp;
-    scan.header.frame_id = "laser";
-
-    auto transform_base = geometry_msgs::msg::TransformStamped{};
-    transform_base.header.stamp = timestamp;
-    transform_base.header.frame_id = "odom";
-    transform_base.child_frame_id = "base_footprint";
-
-    auto transform_laser = geometry_msgs::msg::TransformStamped{};
-    transform_laser.header.stamp = timestamp;
-    transform_laser.header.frame_id = "base_footprint";
-    transform_laser.child_frame_id = "laser";
-
-    laser_scan_publisher_->publish(scan);
-    tf_broadcaster_->sendTransform(transform_base);
-    tf_broadcaster_->sendTransform(transform_laser);
-  }
-
-  void publish_laser_scan_with_no_odom_to_base() {
-    const auto timestamp = now();
-
-    auto scan = sensor_msgs::msg::LaserScan{};
-    scan.header.stamp = timestamp;
-    scan.header.frame_id = "laser";
-
-    auto transform_base = geometry_msgs::msg::TransformStamped{};
-    transform_base.header.stamp = timestamp;
-    transform_base.header.frame_id = "odom";
-    transform_base.child_frame_id = "unexpected_base";
-
-    auto transform_laser = geometry_msgs::msg::TransformStamped{};
-    transform_laser.header.stamp = timestamp;
-    transform_laser.header.frame_id = "unexpected_base";
-    transform_laser.child_frame_id = "laser";
-
-    laser_scan_publisher_->publish(scan);
-    tf_broadcaster_->sendTransform(transform_base);
-    tf_broadcaster_->sendTransform(transform_laser);
-  }
-
-  void publish_laser_scan_with_odom_to_base(const Sophus::SE2d& transform) {
-    const auto timestamp = now();
-
-    auto scan = sensor_msgs::msg::LaserScan{};
-    scan.header.stamp = timestamp;
-    scan.header.frame_id = "laser";
-
-    auto transform_base = geometry_msgs::msg::TransformStamped{};
-    transform_base.header.stamp = timestamp;
-    transform_base.header.frame_id = "odom";
-    transform_base.child_frame_id = "base_footprint";
-    transform_base.transform = tf2::toMsg(transform);
-
-    auto transform_laser = geometry_msgs::msg::TransformStamped{};
-    transform_laser.header.stamp = timestamp;
-    transform_laser.header.frame_id = "base_footprint";
-    transform_laser.child_frame_id = "laser";
-
-    laser_scan_publisher_->publish(scan);
-    tf_broadcaster_->sendTransform(transform_base);
-    tf_broadcaster_->sendTransform(transform_laser);
-  }
-
-  bool can_transform(const std::string& target, const std::string& source) const {
-    return tf_buffer_ && tf_buffer_->canTransform(target, source, tf2::TimePointZero);
-  }
-
-  auto lookup_transform(const std::string& target, const std::string& source) const {
-    auto transform = Sophus::SE2d{};
-    if (tf_buffer_) {
-      tf2::convert(tf_buffer_->lookupTransform(target, source, tf2::TimePointZero).transform, transform);
-    }
-    return transform;
-  }
-
- private:
-  template <class Message>
-  using PublisherPtr = std::shared_ptr<rclcpp::Publisher<Message>>;
-
-  PublisherPtr<geometry_msgs::msg::PoseWithCovarianceStamped> initial_pose_publisher_;
-  PublisherPtr<sensor_msgs::msg::LaserScan> laser_scan_publisher_;
-
-  template <class Message>
-  using SubscriberPtr = std::shared_ptr<rclcpp::Subscription<Message>>;
-
-  SubscriberPtr<geometry_msgs::msg::PoseWithCovarianceStamped> pose_subscriber_;
-  SubscriberPtr<nav2_msgs::msg::ParticleCloud> particle_cloud_subscriber_;
-
-  std::optional<geometry_msgs::msg::PoseWithCovarianceStamped> latest_pose_;
-  std::optional<nav2_msgs::msg::ParticleCloud> latest_particle_cloud_;
-
-  std::unique_ptr<tf2_ros::Buffer> tf_buffer_;
-  std::unique_ptr<tf2_ros::TransformBroadcaster> tf_broadcaster_;
-  std::unique_ptr<tf2_ros::TransformListener> tf_listener_;
-};
-
 /// Base node fixture class with common utilities.
 template <class T>
 class BaseNodeFixture : public T {
@@ -265,7 +54,7 @@ class BaseNodeFixture : public T {
   void SetUp() override {
     rclcpp::init(0, nullptr);
     ndt_amcl_node_ = std::make_shared<AmclNodeUnderTest>();
-    tester_node_ = std::make_shared<TesterNode>();
+    tester_node_ = std::make_shared<beluga_amcl::testing::TesterNode>();
     tester_node_->create_transform_buffer();
     ndt_amcl_node_->set_parameter(rclcpp::Parameter("map_path", kTestMapPath));
   }
@@ -290,7 +79,7 @@ class BaseNodeFixture : public T {
 
  protected:
   std::shared_ptr<AmclNodeUnderTest> ndt_amcl_node_;
-  std::shared_ptr<TesterNode> tester_node_;
+  std::shared_ptr<beluga_amcl::testing::TesterNode> tester_node_;
 };
 
 class TestLifecycle : public BaseNodeFixture<::testing::Test> {};
